requirements.txt
scikit-learn == 1.7.0
pandas == 2.3.0
fastapi == 0.115.12
uvicorn == 0.34.3
Jinja2 == 3.1.6
python-multipart == 0.0.20
psycopg2-binary == 2.9.9
python-dotenv == 1.0.1
